+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## [1.26.1] - 2026-08-30
 
 ### Added
 - **Dundun's Abundant Travel Method** The active Abundance event icon on the world map carries a button to use the toy, teleporting you straight to the event.